[TF] Enable GA Cloud Scheduler
