+++ google-beta/resource_cloud_scheduler_job_generated_test.go
@@ -33,11 +33,17 @@ func TestAccCloudSchedulerJob_schedulerJobPubsubExample(t *testing.T) {
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProvidersOiCS,
+		Providers:    testAccProviders,
 		CheckDestroy: testAccCheckCloudSchedulerJobDestroy,
 		Steps: []resource.TestStep{
 			{
 				Config: testAccCloudSchedulerJob_schedulerJobPubsubExample(context),
+			},
+			{
+				ResourceName:            "google_cloud_scheduler_job.job",
+				ImportState:             true,
+				ImportStateVerify:       true,
+				ImportStateVerifyIgnore: []string{"region"},
 			},
 		},
 	})
@@ -46,14 +52,10 @@ func TestAccCloudSchedulerJob_schedulerJobPubsubExample(t *testing.T) {
 func testAccCloudSchedulerJob_schedulerJobPubsubExample(context map[string]interface{}) string {
 	return Nprintf(`
 resource "google_pubsub_topic" "topic" {
-  provider = "google-beta"
-
   name = "job-topic-%{random_suffix}"
 }
 
 resource "google_cloud_scheduler_job" "job" {
-  provider = "google-beta"
-
   name     = "test-job-%{random_suffix}"
   description = "test job"
   schedule = "*/2 * * * *"
@@ -62,11 +64,6 @@ resource "google_cloud_scheduler_job" "job" {
     topic_name = "${google_pubsub_topic.topic.id}"
     data = "${base64encode("test")}"
   }
-}
-
-provider "google-beta"{
-  region = "us-central1"
-  zone   = "us-central1-a"
 }
 `, context)
 }
@@ -80,11 +77,17 @@ func TestAccCloudSchedulerJob_schedulerJobHttpExample(t *testing.T) {
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProvidersOiCS,
+		Providers:    testAccProviders,
 		CheckDestroy: testAccCheckCloudSchedulerJobDestroy,
 		Steps: []resource.TestStep{
 			{
 				Config: testAccCloudSchedulerJob_schedulerJobHttpExample(context),
+			},
+			{
+				ResourceName:            "google_cloud_scheduler_job.job",
+				ImportState:             true,
+				ImportStateVerify:       true,
+				ImportStateVerifyIgnore: []string{"region"},
 			},
 		},
 	})
@@ -93,8 +96,6 @@ func TestAccCloudSchedulerJob_schedulerJobHttpExample(t *testing.T) {
 func testAccCloudSchedulerJob_schedulerJobHttpExample(context map[string]interface{}) string {
 	return Nprintf(`
 resource "google_cloud_scheduler_job" "job" {
-  provider = "google-beta"
-
   name     = "test-job-%{random_suffix}"
   description = "test http job"
   schedule = "*/8 * * * *"
@@ -104,11 +105,6 @@ resource "google_cloud_scheduler_job" "job" {
     http_method = "POST"
     uri = "https://example.com/ping"
   }
-}
-
-provider "google-beta"{
-  region = "us-central1"
-  zone   = "us-central1-a"
 }
 `, context)
 }
@@ -122,11 +118,17 @@ func TestAccCloudSchedulerJob_schedulerJobAppEngineExample(t *testing.T) {
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:     func() { testAccPreCheck(t) },
-		Providers:    testAccProvidersOiCS,
+		Providers:    testAccProviders,
 		CheckDestroy: testAccCheckCloudSchedulerJobDestroy,
 		Steps: []resource.TestStep{
 			{
 				Config: testAccCloudSchedulerJob_schedulerJobAppEngineExample(context),
+			},
+			{
+				ResourceName:            "google_cloud_scheduler_job.job",
+				ImportState:             true,
+				ImportStateVerify:       true,
+				ImportStateVerifyIgnore: []string{"region"},
 			},
 		},
 	})
@@ -135,8 +137,6 @@ func TestAccCloudSchedulerJob_schedulerJobAppEngineExample(t *testing.T) {
 func testAccCloudSchedulerJob_schedulerJobAppEngineExample(context map[string]interface{}) string {
 	return Nprintf(`
 resource "google_cloud_scheduler_job" "job" {
-  provider = "google-beta"
-
   name     = "test-job-%{random_suffix}"
   schedule = "*/4 * * * *"
   description = "test app engine job"
@@ -154,11 +154,6 @@ resource "google_cloud_scheduler_job" "job" {
     relative_uri = "/ping"
   }
 }
-
-provider "google-beta"{
-  region = "us-central1"
-  zone   = "us-central1-a"
-}
 `, context)
 }
 
@@ -173,7 +168,7 @@ func testAccCheckCloudSchedulerJobDestroy(s *terraform.State) error {
 
 		config := testAccProvider.Meta().(*Config)
 
-		url, err := replaceVarsForTest(rs, "https://cloudscheduler.googleapis.com/v1beta1/projects/{{project}}/locations/{{region}}/jobs/{{name}}")
+		url, err := replaceVarsForTest(rs, "https://cloudscheduler.googleapis.com/v1/projects/{{project}}/locations/{{region}}/jobs/{{name}}")
 		if err != nil {
 			return err
 		}
