+++ google-beta/resource_tpu_node.go
@@ -18,6 +18,7 @@ import (
 	"fmt"
 	"log"
 	"reflect"
+	"regexp"
 	"strconv"
 	"time"
 
@@ -36,6 +37,25 @@ func compareTpuNodeSchedulingConfig(k, old, new string, d *schema.ResourceData) 
 		return o.(bool) == n.(bool)
 	}
 	return false
+}
+
+func validateHttpHeaders() schema.SchemaValidateFunc {
+	return func(i interface{}, k string) (s []string, es []error) {
+		headers := i.(map[string]interface{})
+		if _, ok := headers["Content-Length"]; ok {
+			es = append(es, fmt.Errorf("Cannot set the Content-Length header on %s", k))
+			return
+		}
+		for key := range headers {
+			match, _ := regexp.MatchString("(X-Google-|X-AppEngine-).*", key)
+			if match {
+				es = append(es, fmt.Errorf("Cannot set the %s header on %s", key, k))
+				return
+			}
+		}
+
+		return
+	}
 }
 
 func resourceTpuNode() *schema.Resource {
